#17 ワンクリックショートカット起動
Start login and auto load for launcher without UI control
