+++ KsGameLauncher/Forms/LicensesForm.cs
@@ -17,7 +17,7 @@ namespace KsGameLauncher
 
         private void Licenses_Load(object sender, EventArgs e)
         {
-            Icon = Properties.Resources.app;
+            Icon = Properties.Resources.appIcon;
             Text = Resources.LicenseWindowTitle;
             textBox_LicenseText.Text = Properties.Resources.Licenses;
             button_Close.Text = Resources.ButtonClose;
