+++ KsGameLauncher/KsGameLauncher.csproj
@@ -146,6 +146,15 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="MainContext.cs" />
+    <Compile Include="MainContext.Designer.cs">
+      <DependentUpon>MainContext.cs</DependentUpon>
+    </Compile>
+    <Compile Include="MainContext.resx.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>MainContext.resx</DependentUpon>
+    </Compile>
     <Compile Include="NotifyIconContextMenuStrip.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -231,6 +240,18 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\OptionsForm.resx">
       <DependentUpon>OptionsForm.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="MainContext.Designer.cs">
+      <DependentUpon>MainContext.cs</DependentUpon>
+      <SubType>Designer</SubType>
+    </EmbeddedResource>
+    <EmbeddedResource Include="MainContext.ja-JP.resx">
+      <DependentUpon>MainContext.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="MainContext.resx">
+      <DependentUpon>MainContext.cs</DependentUpon>
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>MainContext.resx.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
