+++ KsGameLauncher/MainContext.ja-JP.resx
@@ -117,53 +117,19 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="appIcon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\resources\icons\app.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  <data name="aboutToolStripMenuItem.Text" xml:space="preserve">
+    <value>アプリについて(&amp;A)</value>
   </data>
-  <data name="AppIconDesigner" xml:space="preserve">
-    <value>anon</value>
+  <data name="addNewGameToolStripMenuItem.Text" xml:space="preserve">
+    <value>ゲームを追加(&amp;N)</value>
   </data>
-  <data name="AuthorizeDomain" xml:space="preserve">
-    <value>account.konami.net</value>
+  <data name="exitToolStripMenuItem.Text" xml:space="preserve">
+    <value>終了(&amp;X)</value>
   </data>
-  <data name="CredentialTarget" xml:space="preserve">
-    <value>Ks Game Launcher</value>
+  <data name="manageAccountsToolStripMenuItem.Text" xml:space="preserve">
+    <value>アカウント管理(&amp;M)</value>
   </data>
-  <data name="default_game_icon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\resources\icons\default-game-icon.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Developers" xml:space="preserve">
-    <value>anon</value>
-  </data>
-  <data name="icon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\resources\icons\icon256.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Licenses" xml:space="preserve">
-    <value>AdysTech.CredentialManager (https://github.com/AdysTech/CredentialManager) 
-
-HtmlAgilityPack (https://github.com/atifaziz/Hazz)</value>
-    <comment>Using libraries</comment>
-  </data>
-  <data name="LoginCookieSessionKey" xml:space="preserve">
-    <value>M573SSID</value>
-    <comment>Login cookie key</comment>
-  </data>
-  <data name="MaintenanceCheckString" xml:space="preserve">
-    <value>メンテナンス期間</value>
-    <comment>Detect words under maintenance</comment>
-  </data>
-  <data name="SpecialThanks" xml:space="preserve">
-    <value>KONAMI Amusement Inc.
-flaticon ( https://www.flaticon.com )</value>
-  </data>
-  <data name="SupportLabelText" xml:space="preserve">
-    <value>Support</value>
-  </data>
-  <data name="SupportLabelURL" xml:space="preserve">
-    <value>https://github.com/anon5r/KsGameLauncher</value>
-  </data>
-  <data name="TosCheckPath" xml:space="preserve">
-    <value>/terms_of_service/index.html</value>
+  <data name="optionsToolStripMenuItem.Text" xml:space="preserve">
+    <value>オプション(&amp;O)</value>
   </data>
 </root>
