+++ KsGameLauncher/MainContext.resx
@@ -117,53 +117,61 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="appIcon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\resources\icons\app.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  <data name="&gt;&gt;aboutToolStripMenuItem.Name" xml:space="preserve">
+    <value>aboutToolStripMenuItem</value>
   </data>
-  <data name="AppIconDesigner" xml:space="preserve">
-    <value>anon</value>
+  <data name="&gt;&gt;aboutToolStripMenuItem.Type" xml:space="preserve">
+    <value>System.Windows.Forms.ToolStripMenuItem, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </data>
-  <data name="AuthorizeDomain" xml:space="preserve">
-    <value>account.konami.net</value>
+  <data name="&gt;&gt;addNewGameToolStripMenuItem.Name" xml:space="preserve">
+    <value>addNewGameToolStripMenuItem</value>
   </data>
-  <data name="CredentialTarget" xml:space="preserve">
-    <value>Ks Game Launcher</value>
+  <data name="&gt;&gt;addNewGameToolStripMenuItem.Type" xml:space="preserve">
+    <value>System.Windows.Forms.ToolStripMenuItem, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </data>
-  <data name="default_game_icon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\resources\icons\default-game-icon.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  <data name="&gt;&gt;contextMenuStrip_Sub.Name" xml:space="preserve">
+    <value>contextMenuStrip_Sub</value>
   </data>
-  <data name="Developers" xml:space="preserve">
-    <value>anon</value>
+  <data name="&gt;&gt;contextMenuStrip_Sub.Type" xml:space="preserve">
+    <value>KsGameLauncher.NotifyIconContextMenuStrip, ksgamelauncher, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</value>
   </data>
-  <data name="icon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\resources\icons\icon256.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  <data name="&gt;&gt;exitToolStripMenuItem.Name" xml:space="preserve">
+    <value>exitToolStripMenuItem</value>
   </data>
-  <data name="Licenses" xml:space="preserve">
-    <value>AdysTech.CredentialManager (https://github.com/AdysTech/CredentialManager) 
-
-HtmlAgilityPack (https://github.com/atifaziz/Hazz)</value>
-    <comment>Using libraries</comment>
+  <data name="&gt;&gt;exitToolStripMenuItem.Type" xml:space="preserve">
+    <value>System.Windows.Forms.ToolStripMenuItem, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </data>
-  <data name="LoginCookieSessionKey" xml:space="preserve">
-    <value>M573SSID</value>
-    <comment>Login cookie key</comment>
+  <data name="&gt;&gt;manageAccountsToolStripMenuItem.Name" xml:space="preserve">
+    <value>manageAccountsToolStripMenuItem</value>
   </data>
-  <data name="MaintenanceCheckString" xml:space="preserve">
-    <value>メンテナンス期間</value>
-    <comment>Detect words under maintenance</comment>
+  <data name="&gt;&gt;manageAccountsToolStripMenuItem.Type" xml:space="preserve">
+    <value>System.Windows.Forms.ToolStripMenuItem, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </data>
-  <data name="SpecialThanks" xml:space="preserve">
-    <value>KONAMI Amusement Inc.
-flaticon ( https://www.flaticon.com )</value>
+  <data name="&gt;&gt;optionsToolStripMenuItem.Name" xml:space="preserve">
+    <value>optionsToolStripMenuItem</value>
   </data>
-  <data name="SupportLabelText" xml:space="preserve">
-    <value>Support</value>
+  <data name="&gt;&gt;optionsToolStripMenuItem.Type" xml:space="preserve">
+    <value>System.Windows.Forms.ToolStripMenuItem, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </data>
-  <data name="SupportLabelURL" xml:space="preserve">
-    <value>https://github.com/anon5r/KsGameLauncher</value>
+  <data name="&gt;&gt;toolStripSeparator_Bottom.Name" xml:space="preserve">
+    <value>toolStripSeparator_Bottom</value>
   </data>
-  <data name="TosCheckPath" xml:space="preserve">
-    <value>/terms_of_service/index.html</value>
+  <data name="&gt;&gt;toolStripSeparator_Bottom.Type" xml:space="preserve">
+    <value>System.Windows.Forms.ToolStripSeparator, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </data>
+  <data name="aboutToolStripMenuItem.Text" xml:space="preserve">
+    <value>&amp;About</value>
+  </data>
+  <data name="addNewGameToolStripMenuItem.Text" xml:space="preserve">
+    <value>Add &amp;new game</value>
+  </data>
+  <data name="exitToolStripMenuItem.Text" xml:space="preserve">
+    <value>E&amp;xit</value>
+  </data>
+  <data name="manageAccountsToolStripMenuItem.Text" xml:space="preserve">
+    <value>&amp;Manage account</value>
+  </data>
+  <data name="optionsToolStripMenuItem.Text" xml:space="preserve">
+    <value>&amp;Options</value>
   </data>
 </root>
